+++ demo/pages/elements/calendar/CalendarDemo.ts
@@ -4,13 +4,10 @@ import { CalendarEvent, CalendarEventResponse } from './../../../../index';
 // APP
 let BigCalendarDemoTpl = require('./templates/BigCalendarDemo.html');
 let CalendarDemoTpl = require('./templates/CalendarDemo.html');
-let TimeDemoTpl = require('./templates/TimeDemo.html');
-let RangeDemoTpl = require('./templates/RangeDemo.html');
-let DateTimeDemoTpl = require('./templates/DateTimeDemo.html');
 
 const template = `
 <div class="container">
-    <h1>Date and Time Pickers</h1>
+    <h1>Calendars & Schedules</h1>
     <p>These allow users to easily select a time and date. It comes in a handful of varieties based on the content of the field.</p>
 
     <h2>Calendar Picker  <small><a target="_blank" href="https://github.com/bullhorn/novo-elements/blob/master/src/elements/date-picker">(source)</a></small></h2>
@@ -23,25 +20,13 @@ const template = `
     <h2>Time Picker  <small><a target="_blank" href="https://github.com/bullhorn/novo-elements/blob/master/src/elements/time-picker">(source)</a></small></h2>
     <p>Time pickers come in 12 hour or 24 hour style.</p>
 
-    <h5>Standalone Time Picker</h5>
-    <div class="example demo">${TimeDemoTpl}</div>
-    <code-snippet [code]="TimeDemoTpl"></code-snippet>
-
-    <h5>Range Picker</h5>
-    <div class="example demo">${RangeDemoTpl}</div>
-    <code-snippet [code]="RangeDemoTpl"></code-snippet>
-
-    <h5>Date Time Picker</h5>
-    <div class="example demo">${DateTimeDemoTpl}</div>
-    <code-snippet [code]="DateTimeDemoTpl"></code-snippet>
-
     <h5>Big Calendar Picker</h5>
     <div class="example demo">${BigCalendarDemoTpl}</div>
     <code-snippet [code]="BigCalendarDemoTpl"></code-snippet>
 </div>
 `;
 
-export const colors: any = {
+const colors: any = {
     red: {
         primary: '#ad2121',
         secondary: '#FAE3E3'
@@ -67,35 +52,54 @@ export const colors: any = {
 export class CalendarDemoComponent {
     BigCalendarDemoTpl: string = BigCalendarDemoTpl;
     CalendarDemoTpl: string = CalendarDemoTpl;
-    TimeDemoTpl: string = TimeDemoTpl;
-    RangeDemoTpl: string = RangeDemoTpl;
-    DateTimeDemoTpl: string = DateTimeDemoTpl;
-
-    time: Date = new Date();
-    dateOne: Date = new Date();
-    dateTwo: Date = new Date();
-    dateTime: Date = new Date();
-    start: any = new Date().setMonth(new Date().getMonth() - 1);
-    end: any = new Date().setMonth(new Date().getMonth() + 1);
-    value: any = {
-        startDate: null,
-        endDate: null
-    };
+    private views:Array<any> = [
+        {
+            label: 'Month',
+            value: 'month'
+        },
+        {
+            label: 'Week',
+            value: 'week'
+        },
+        {
+            label: 'Day',
+            value: 'day'
+        }
+    ];
+    view:string = 'month';
 
     viewDate: Date = new Date();
     events: CalendarEvent[] = [{
-        title: 'Has custom class',
-        color: colors.red,
+        title: 'Interview',
+        description: 'with @bvkimball',
+        color: colors.green,
         start: new Date(),
-        response: CalendarEventResponse.Rejected
+        response: CalendarEventResponse.Rejected,
+        type: 'Interview'
+    }, {
+        title: 'Client Visit',
+        description: 'with @asibilia',
+        color: colors.red,
+        start: new Date(Date.now() + (60000 * 30)),
+        response: CalendarEventResponse.Accepted,
+        type: 'Visit'
+    }, {
+        title: 'Interview',
+        description: 'with @johndoe',
+        color: colors.green,
+        start: new Date(Date.now() + (60000 * 90)),
+        response: CalendarEventResponse.Accepted,
+        type: 'Interview'
     }];
 
     getNewEvent(date, color, type): CalendarEvent {
         let evt: CalendarEvent = {
-            title: 'Has custom class',
+            title: 'Meeting',
+            description: 'with @jgodi',
             color: color,
             start: date,
-            response: type
+            response: type,
+            type: 'Meeting'
         };
         return evt;
     }
@@ -103,11 +107,13 @@ export class CalendarDemoComponent {
     dayClicked(event) {
         let evt: CalendarEvent = this.getNewEvent(event.day.date, colors.blue, CalendarEventResponse.Maybe);
         this.events.push(evt);
+        this.events = [...this.events];
     }
 
     addShift(event) {
         let evt: CalendarEvent = this.getNewEvent(event.day.date, colors.blue, CalendarEventResponse.Maybe);
         this.events.push(evt);
+        this.events = [...this.events];
     }
 
     removeShift(event) {
@@ -119,6 +125,7 @@ export class CalendarDemoComponent {
         if (!event.day.events.length) {
             evt = this.getNewEvent(event.day.date, colors.green, CalendarEventResponse.Accepted);
             this.events.push(evt);
+            this.events = [...this.events];
         } else {
             evt = event.day.events[0];
             switch (evt.response) {