+++ src/elements/header/Header.ts
@@ -26,11 +26,23 @@ export class UtilActionElement {
     selector: 'header[theme]',
     host: {
         '[class]': 'headerClass',
-        '[attr.theme]': 'theme'
+        '[attr.theme]': 'theme',
+        '[class.hasSubTitle]': 'subTitle'
     },
     template: `
         <section class="primary" *ngIf="theme || config.theme">
-            <h1><i *ngIf="icon" [ngClass]="iconClass"></i>{{title || config.title}}</h1>
+            <h1>
+                <i *ngIf="icon" [ngClass]="iconClass"></i>
+                <ng-container *ngIf="!subTitle && !config.subTitle">
+                    {{ title || config.title }}
+                </ng-container>
+                <ng-container *ngIf="subTitle || config.subTitle">
+                    <span>
+                        {{ title || config.title }}
+                        <small *ngIf="subTitle">{{ subTitle || config.subTitle }}</small>
+                    </span>
+                </ng-container>
+            </h1>
             <ng-content select="utils"></ng-content>
         </section>
         <ng-content></ng-content>
@@ -38,6 +50,7 @@ export class UtilActionElement {
 })
 export class NovoHeaderElement implements OnInit {
     @Input() title: string;
+    @Input() subTitle: string;
     @Input() theme: string;
     @Input() icon: string;
     @Input() config: any;